Refactor mock call asserter
