+++ classes/asserters/mock.php
@@ -7,18 +7,39 @@ use
 	mageekguy\atoum\php,
 	mageekguy\atoum\test,
 	mageekguy\atoum\asserter,
-	mageekguy\atoum\exceptions,
-	mageekguy\atoum\tools\arguments
+	mageekguy\atoum\exceptions
 ;
 
 class mock extends atoum\asserter
 {
-	protected $mock = null;
-	protected $call = null;
-	protected $beforeFunctionCalls = array();
-	protected $afterFunctionCalls = array();
-	protected $beforeMethodCalls = array();
-	protected $afterMethodCalls = array();
+	protected $mock;
+	protected $callAsserter;
+
+	public function __call($method, $arguments)
+	{
+		switch ($method)
+		{
+			case 'beforeMethodCall':
+			case 'afterMethodCall':
+				return call_user_func_array(array($this->getCallAsserter(), $method), array($arguments[0], $this->mock));
+
+			case 'withArguments':
+			case 'withIdenticalArguments':
+			case 'withAnyArguments':
+			case 'withoutAnyArgument':
+			case 'withAtLeastArguments':
+			case 'withAnyMethodCallsBefore':
+			case 'withAnyMethodCallsAfter':
+			case 'withAnyFunctionCallsBefore':
+			case 'withAnyFunctionCallsAfter':
+			case 'beforeFunctionCall':
+			case 'afterFunctionCall':
+				return call_user_func_array(array($this->getCallAsserter(), $method), $arguments);
+
+			default:
+				return parent::__call($method, $arguments);
+		}
+	}
 
 	public function reset()
 	{
@@ -28,11 +49,6 @@ class mock extends atoum\asserter
 		}
 
 		return $this;
-	}
-
-	public function getCall()
-	{
-		return ($this->call === null ? null : clone $this->call);
 	}
 
 	public function setWith($mock)
@@ -54,82 +70,6 @@ class mock extends atoum\asserter
 	public function getMock()
 	{
 		return $this->mock;
-	}
-
-	public function beforeMethodCall($methodName)
-	{
-		$this->mockIsSet()->beforeMethodCalls[] = $beforeMethodCall = new mock\call\mock($this, $this->mock, $methodName);
-
-		return $beforeMethodCall;
-	}
-
-	public function getBeforeMethodCalls()
-	{
-		return $this->beforeMethodCalls;
-	}
-
-	public function withAnyMethodCallsBefore()
-	{
-		$this->beforeMethodCalls = array();
-
-		return $this;
-	}
-
-	public function afterMethodCall($methodName)
-	{
-		$this->mockIsSet()->afterMethodCalls[] = $afterMethodCall = new mock\call\mock($this, $this->mock, $methodName);
-
-		return $afterMethodCall;
-	}
-
-	public function getAfterMethodCalls()
-	{
-		return $this->afterMethodCalls;
-	}
-
-	public function withAnyMethodCallsAfter()
-	{
-		$this->afterMethodCalls = array();
-
-		return $this;
-	}
-
-	public function beforeFunctionCall($functionName, test\adapter $adapter)
-	{
-		$this->mockIsSet()->beforeFunctionCalls[] = $beforeFunctionCall = new mock\call\adapter($this, $adapter, $functionName);
-
-		return $beforeFunctionCall;
-	}
-
-	public function getBeforeFunctionCalls()
-	{
-		return $this->beforeFunctionCalls;
-	}
-
-	public function withAnyFunctionCallsBefore()
-	{
-		$this->beforeFunctionCalls = array();
-
-		return $this;
-	}
-
-	public function afterFunctionCall($functionName, test\adapter $adapter)
-	{
-		$this->mockIsSet()->afterFunctionCalls[] = $afterFunctionCall = new mock\call\adapter($this, $adapter, $functionName);
-
-		return $afterFunctionCall;
-	}
-
-	public function getAfterFunctionCalls()
-	{
-		return $this->afterFunctionCalls;
-	}
-
-	public function withAnyFunctionCallsAfter()
-	{
-		$this->afterFunctionCalls = array();
-
-		return $this;
 	}
 
 	public function wasCalled($failMessage = null)
@@ -160,126 +100,19 @@ class mock extends atoum\asserter
 		return $this;
 	}
 
+	public function getCallAsserter()
+	{
+		if ($this->mockIsSet()->callAsserter === null)
+		{
+			$this->callAsserter = new call\mock($this);
+		}
+
+		return $this->callAsserter;
+	}
+
 	public function call($function)
 	{
-		if ($this->mockIsSet()->call === null)
-		{
-			$this->call = new php\call($function, null, $this->mock);
-		}
-		else
-		{
-			$this->call
-				->setFunction($function)
-				->setObject($this->mock)
-				->unsetArguments()
-			;
-		}
-
-		return $this;
-	}
-
-	public function withArguments()
-	{
-		$this->calledMethodNameIsSet()->call->setArguments(func_get_args())->notIdentical();
-
-		return $this;
-	}
-
-	public function withIdenticalArguments()
-	{
-		$this->calledMethodNameIsSet()->call->setArguments(func_get_args())->identical();
-
-		return $this;
-	}
-
-	public function withAtLeastArguments(array $arguments)
-	{
-		$this->calledMethodNameIsSet()->call->setArguments($arguments)->notIdentical();
-
-		return $this;
-	}
-
-	public function withAtLeastIdenticalArguments(array $arguments)
-	{
-		$this->calledMethodNameIsSet()->call->setArguments($arguments)->identical();
-
-		return $this;
-	}
-
-	public function withAnyArguments()
-	{
-		$this->calledMethodNameIsSet()->call->unsetArguments();
-
-		return $this;
-	}
-
-	public function withoutAnyArgument()
-	{
-		$this->calledMethodNameIsSet()->call->setArguments(array());
-
-		return $this;
-	}
-
-	public function once($failMessage = null)
-	{
-		return $this->exactly(1, $failMessage);
-	}
-
-	public function twice($failMessage = null)
-	{
-		return $this->exactly(2, $failMessage);
-	}
-
-	public function thrice($failMessage = null)
-	{
-		return $this->exactly(3, $failMessage);
-	}
-
-	public function atLeastOnce($failMessage = null)
-	{
-		$calls = $this->assertOnBeforeAndAfterCalls();
-
-		if (($callsNumber = sizeof($calls)) >= 1)
-		{
-			$this->pass();
-		}
-		else
-		{
-			$this->fail($failMessage !== null ? $failMessage : sprintf($this->getLocale()->_('method %s is called 0 time'), $this->call) . $this->getCallsAsString());
-		}
-
-		return $this;
-	}
-
-	public function exactly($number, $failMessage = null)
-	{
-		$calls = $this->assertOnBeforeAndAfterCalls();
-
-		if (($callsNumber = sizeof($calls)) == $number)
-		{
-			$this->pass();
-		}
-		else
-		{
-			$this->fail($failMessage !== null ? $failMessage : sprintf(
-					$this->getLocale()->__(
-						'method %s is called %d time instead of %d',
-						'method %s is called %d times instead of %d',
-						$callsNumber
-					),
-					$this->call,
-					$callsNumber,
-					$number
-				) . $this->getCallsAsString()
-			);
-		}
-
-		return $this;
-	}
-
-	public function never($failMessage = null)
-	{
-		return $this->exactly(0, $failMessage);
+		return $this->getCallAsserter()->setWith(new php\call($function, null, $this->mock));
 	}
 
 	protected function mockIsSet()
@@ -290,119 +123,5 @@ class mock extends atoum\asserter
 		}
 
 		return $this;
-	}
-
-	protected function calledMethodNameIsSet()
-	{
-		if ($this->mockIsSet()->call === null)
-		{
-			throw new exceptions\logic('Called method is undefined');
-		}
-
-		return $this;
-	}
-
-	protected function assertOnBeforeAndAfterCalls()
-	{
-		$calls = $this->calledMethodNameIsSet()->mock->getMockController()->getCalls($this->call->getFunction(), $this->call->getArguments(), $this->call->isIdentical());
-
-		if (sizeof($calls) > 0)
-		{
-			foreach ($this->beforeMethodCalls as $beforeMethodCall)
-			{
-				$firstCall = $beforeMethodCall->getFirstCall();
-
-				if ($firstCall === null)
-				{
-					$this->fail(sprintf($this->getLocale()->_('method %s is not called'), $beforeMethodCall));
-				}
-
-				if (key($calls) > $firstCall)
-				{
-					$this->fail(sprintf($this->getLocale()->_('method %s is not called before method %s'), $this->call, $beforeMethodCall));
-				}
-
-				$this->pass();
-			}
-
-			foreach ($this->beforeFunctionCalls as $beforeFunctionCall)
-			{
-				$firstCall = $beforeFunctionCall->getFirstCall();
-
-				if ($firstCall === null)
-				{
-					$this->fail(sprintf($this->getLocale()->_('method %s is not called'), $beforeFunctionCall));
-				}
-
-				if (key($calls) > $firstCall)
-				{
-					$this->fail(sprintf($this->getLocale()->_('method %s is not called before function %s'), $this->call, $beforeFunctionCall));
-				}
-
-				$this->pass();
-			}
-
-			foreach ($this->afterMethodCalls as $afterMethodCall)
-			{
-				$lastCall = $afterMethodCall->getLastCall();
-
-				if ($lastCall === null)
-				{
-					$this->fail(sprintf($this->getLocale()->_('method %s is not called'), $afterMethodCall));
-				}
-
-				if (key($calls) < $lastCall)
-				{
-					$this->fail(sprintf($this->getLocale()->_('method %s is not called after method %s'), $this->call, $afterMethodCall));
-				}
-
-				$this->pass();
-			}
-
-			foreach ($this->afterFunctionCalls as $afterFunctionCall)
-			{
-				$lastCall = $afterFunctionCall->getLastCall();
-
-				if ($lastCall === null)
-				{
-					$this->fail(sprintf($this->getLocale()->_('method %s is not called'), $afterFunctionCall));
-				}
-
-				if (key($calls) < $lastCall)
-				{
-					$this->fail(sprintf($this->getLocale()->_('method %s is not called after function %s'), $this->call, $afterFunctionCall));
-				}
-
-				$this->pass();
-			}
-		}
-
-		$this->beforeMethodCalls = array();
-		$this->afterMethodCalls = array();
-		$this->beforeFunctionCalls = array();
-		$this->afterFunctionCalls = array();
-
-		return $calls;
-	}
-
-	protected function getCallsAsString()
-	{
-		$string = '';
-
-		if (($calls  = $this->mock->getMockController()->getCalls($this->call->getFunction())) !== null)
-		{
-			$format = '[%' . strlen((string) sizeof($calls)) . 's] %s';
-
-			$phpCalls = array();
-
-			foreach (array_values($calls) as $call => $arguments)
-			{
-				$phpCalls[] = sprintf($format, $call + 1, new php\call($this->call->getFunction(), $arguments, $this->mock));
-			}
-
-			$string = PHP_EOL . join(PHP_EOL, $phpCalls);
-		}
-
-		return $string;
 	}
 }
